Refactor megakernel device ops into common/ package (one module per op+dtype)
