+++ examples/33_gpt_oss_megakernel/gpt_oss_120b_quantized_megakernel.py
@@ -46,643 +46,15 @@ from reference import GptOssConfig, build_yarn_rope
 from load_hf import load_hf_weights
 from tokenizer_util import load_tokenizer
 
+# Device ops live in common/ (one module per op+dtype) so the single-GPU megakernel
+# below and the planned per-GPU multi-GPU kernels share the same kernels.
+from common import *  # noqa: F401,F403  (re-exports every @triton.jit device op)
+
 # Number of persistent programs. Fewer than the 256 CUs on MI355X: with a grid-wide
 # barrier between phases, a smaller grid makes each barrier cheaper while still
 # keeping every phase's GEMV well filled. 180 measured fastest for this model.
 NUM_WG = 180
 _NWG = tl.constexpr(NUM_WG)
-
-
-# ───────────────────────── device helpers ─────────────────────────
-@triton.jit
-def _barrier(bar_ptr, target):
-    # Arrive: release so this program's phase writes are flushed to the shared L2
-    # before the counter increment becomes visible to peers.
-    tl.debug_barrier()
-    tl.atomic_add(bar_ptr, 1, sem="release", scope="gpu")
-    # Spin on a relaxed read: polling with acquire emits a full L1 invalidate
-    # (buffer_inv sc1) every iteration, which dominates the barrier cost. A relaxed
-    # poll just reads the counter cheaply.
-    done = 0
-    while done == 0:
-        cur = tl.atomic_add(bar_ptr, 0, sem="relaxed", scope="gpu")
-        if cur >= target:
-            done = 1
-    # One acquire after the count is reached invalidates L1 a single time, so the
-    # next phase reads every peer's writes fresh from L2.
-    _ = tl.atomic_add(bar_ptr, 0, sem="acquire", scope="gpu")
-
-
-@triton.jit
-def _fp4_lut(mag_idx):
-    return tl.where(
-        mag_idx == 0,
-        0.0,
-        tl.where(
-            mag_idx == 1,
-            0.5,
-            tl.where(
-                mag_idx == 2,
-                1.0,
-                tl.where(
-                    mag_idx == 3,
-                    1.5,
-                    tl.where(mag_idx == 4, 2.0, tl.where(mag_idx == 5, 3.0, tl.where(mag_idx == 6, 4.0, 6.0))),
-                ),
-            ),
-        ),
-    )
-
-
-@triton.jit
-def _gemv_bf16_tiled(
-    w_base,
-    x_ptr,
-    y_ptr,
-    has_bias,
-    b_base,
-    M,
-    K: tl.constexpr,
-    pid,
-    BLOCK_M: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    NSTAGES: tl.constexpr = 3,
-):
-    """y[r] = sum_k w[r,k]*x[k] (+b[r]) for a batch-1 GEMV.
-
-    Each program owns a tile of BLOCK_M contiguous output rows and walks K in
-    BLOCK_K-wide steps, loading a 2D [BLOCK_M, BLOCK_K] weight tile so the inner
-    (contiguous-K) dimension vectorizes to dwordx4. max_contiguous/multiple_of tell
-    the compiler the K index is contiguous and aligned, enabling wide loads. Tiles
-    are strided across the grid by num programs."""
-    npid = tl.num_programs(0)
-    mo = tl.arange(0, BLOCK_M)
-    ko = tl.max_contiguous(tl.multiple_of(tl.arange(0, BLOCK_K), BLOCK_K), BLOCK_K)
-    n_tiles = (M + BLOCK_M - 1) // BLOCK_M
-    NK: tl.constexpr = (K + BLOCK_K - 1) // BLOCK_K
-    tile = pid
-    while tile < n_tiles:
-        rows = tile * BLOCK_M + mo
-        rmask = rows < M
-        acc = tl.zeros((BLOCK_M, BLOCK_K), dtype=tl.float32)
-        for ki in tl.range(0, NK, num_stages=NSTAGES):
-            kk = ki * BLOCK_K + ko
-            kmask = kk < K
-            w = tl.load(w_base + rows[:, None] * K + kk[None, :], mask=rmask[:, None] & kmask[None, :], other=0.0).to(
-                tl.float32
-            )
-            x = tl.load(x_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            acc += w * x[None, :]
-        s = tl.sum(acc, axis=1)  # [BLOCK_M]
-        if has_bias:
-            s += tl.load(b_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        tl.store(y_ptr + rows, s, mask=rmask)
-        tile += npid
-
-
-@triton.jit
-def _store_resid_rmsnorm(
-    x_ptr, o_ptr, g_ptr, out_ptr, H: tl.constexpr, pid, eps, BLOCK_M: tl.constexpr, NORMK: tl.constexpr
-):
-    """Materialize normed = rmsnorm(x + o)*g into out_ptr (bf16), striped across WGs.
-    Each program computes the rms scalar from the full (x + o), then writes its row
-    slices. Used for the non-quantized expert input; needs a barrier after."""
-    noff = tl.arange(0, NORMK)
-    nmask = noff < H
-    xall = tl.load(x_ptr + noff, mask=nmask, other=0.0).to(tl.float32) + tl.load(
-        o_ptr + noff, mask=nmask, other=0.0
-    ).to(tl.float32)
-    ss = tl.sum(xall * xall, axis=0)
-    rms = 1.0 / tl.sqrt(ss / H + eps)
-    base = pid * BLOCK_M
-    step = tl.num_programs(0) * BLOCK_M
-    while base < H:
-        off = base + tl.arange(0, BLOCK_M)
-        m = off < H
-        xv = tl.load(x_ptr + off, mask=m, other=0.0).to(tl.float32) + tl.load(o_ptr + off, mask=m, other=0.0).to(
-            tl.float32
-        )
-        g = tl.load(g_ptr + off, mask=m, other=0.0).to(tl.float32)
-        tl.store(out_ptr + off, (xv * rms * g).to(tl.bfloat16), mask=m)
-        base += step
-
-
-@triton.jit
-def _gemv_bf16_resid_rmsnorm(
-    w_base,
-    x_ptr,
-    o_ptr,
-    g_ptr,
-    y_ptr,
-    has_bias,
-    b_base,
-    M,
-    H: tl.constexpr,
-    pid,
-    eps,
-    BLOCK_M: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    NORMK: tl.constexpr,
-    NSTAGES: tl.constexpr = 3,
-):
-    """Fused residual + RMSNorm + GEMV: y = GEMV(rmsnorm(x + o) * g).
-
-    Reads the attention output o and the residual x and uses (x + o) as the norm
-    input, without writing it back. This folds the post-attention residual add and
-    its RMSNorm (and the grid barrier between them) into the router GEMV. The single
-    residual write x += o happens later in the layer's final accumulation. Each
-    program derives the rms scalar from the full (x + o); both inputs are stable
-    since the previous barrier, so no barrier is needed here."""
-    noff = tl.arange(0, NORMK)
-    nmask = noff < H
-    xall = tl.load(x_ptr + noff, mask=nmask, other=0.0).to(tl.float32) + tl.load(
-        o_ptr + noff, mask=nmask, other=0.0
-    ).to(tl.float32)
-    ss = tl.sum(xall * xall, axis=0)
-    rms = 1.0 / tl.sqrt(ss / H + eps)
-    npid = tl.num_programs(0)
-    mo = tl.arange(0, BLOCK_M)
-    ko = tl.max_contiguous(tl.multiple_of(tl.arange(0, BLOCK_K), BLOCK_K), BLOCK_K)
-    n_tiles = (M + BLOCK_M - 1) // BLOCK_M
-    NK: tl.constexpr = (H + BLOCK_K - 1) // BLOCK_K
-    tile = pid
-    while tile < n_tiles:
-        rows = tile * BLOCK_M + mo
-        rmask = rows < M
-        acc = tl.zeros((BLOCK_M, BLOCK_K), dtype=tl.float32)
-        for ki in tl.range(0, NK, num_stages=NSTAGES):
-            kk = ki * BLOCK_K + ko
-            kmask = kk < H
-            # recompute xnew = x + o inline (avoids depending on other programs'
-            # persisted x_ptr writes -- no barrier needed before this read)
-            xk = tl.load(x_ptr + kk, mask=kmask, other=0.0).to(tl.float32) + tl.load(
-                o_ptr + kk, mask=kmask, other=0.0
-            ).to(tl.float32)
-            gk = tl.load(g_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            nk = xk * rms * gk
-            w = tl.load(w_base + rows[:, None] * H + kk[None, :], mask=rmask[:, None] & kmask[None, :], other=0.0).to(
-                tl.float32
-            )
-            acc += w * nk[None, :]
-        s = tl.sum(acc, axis=1)
-        if has_bias:
-            s += tl.load(b_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        tl.store(y_ptr + rows, s, mask=rmask)
-        tile += npid
-
-
-@triton.jit
-def _gemv_bf16_rmsnorm(
-    w_base,
-    x_ptr,
-    g_ptr,
-    y_ptr,
-    has_bias,
-    b_base,
-    M,
-    H: tl.constexpr,
-    pid,
-    eps,
-    BLOCK_M: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    NORMK: tl.constexpr,
-    NSTAGES: tl.constexpr = 3,
-):
-    """Fused RMSNorm + GEMV: y[r] = sum_k (rmsnorm(x)*g)[k] * w[r,k] (+b[r]).
-
-    Each program computes the RMSNorm scale from x_ptr once (redundant across WGs,
-    but cheap and barrier-free: x is stable since the previous barrier), then runs
-    the tiled GEMV applying norm*gamma inline. Removes the separate pid0 RMSNorm
-    phase AND its grid barrier. NORMK = round_up_pow2(H)."""
-    noff = tl.arange(0, NORMK)
-    nmask = noff < H
-    xall = tl.load(x_ptr + noff, mask=nmask, other=0.0).to(tl.float32)
-    ss = tl.sum(xall * xall, axis=0)
-    rms = 1.0 / tl.sqrt(ss / H + eps)
-    npid = tl.num_programs(0)
-    mo = tl.arange(0, BLOCK_M)
-    ko = tl.max_contiguous(tl.multiple_of(tl.arange(0, BLOCK_K), BLOCK_K), BLOCK_K)
-    n_tiles = (M + BLOCK_M - 1) // BLOCK_M
-    NK: tl.constexpr = (H + BLOCK_K - 1) // BLOCK_K
-    tile = pid
-    while tile < n_tiles:
-        rows = tile * BLOCK_M + mo
-        rmask = rows < M
-        acc = tl.zeros((BLOCK_M, BLOCK_K), dtype=tl.float32)
-        for ki in tl.range(0, NK, num_stages=NSTAGES):
-            kk = ki * BLOCK_K + ko
-            kmask = kk < H
-            xk = tl.load(x_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            gk = tl.load(g_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            nk = xk * rms * gk  # normed activation chunk
-            w = tl.load(w_base + rows[:, None] * H + kk[None, :], mask=rmask[:, None] & kmask[None, :], other=0.0).to(
-                tl.float32
-            )
-            acc += w * nk[None, :]
-        s = tl.sum(acc, axis=1)
-        if has_bias:
-            s += tl.load(b_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        tl.store(y_ptr + rows, s, mask=rmask)
-        tile += npid
-
-
-@triton.jit
-def _gemv_bf16(w_base, x_ptr, y_ptr, has_bias, b_base, M, K: tl.constexpr, pid, BLOCK_K: tl.constexpr):
-    """y[r] = sum_k w[r,k]*x[k] (+b[r]); rows strided across programs by pid. Legacy
-    scalar path kept for reference/fallback; _gemv_bf16_tiled is the fast one."""
-    koff = tl.arange(0, BLOCK_K)
-    r = pid
-    while r < M:
-        acc = tl.zeros((BLOCK_K,), dtype=tl.float32)
-        k0 = 0
-        while k0 < K:
-            kk = k0 + koff
-            kmask = kk < K
-            w = tl.load(w_base + r * K + kk, mask=kmask, other=0.0).to(tl.float32)
-            x = tl.load(x_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            acc += w * x
-            k0 += BLOCK_K
-        s = tl.sum(acc, axis=0)
-        if has_bias:
-            s += tl.load(b_base + r).to(tl.float32)
-        tl.store(y_ptr + r, s)
-        r += _NWG
-
-
-@triton.jit
-def _gemv_fp8_rmsnorm(
-    w_base,
-    ws_base,
-    x_ptr,
-    g_ptr,
-    y_ptr,
-    has_bias,
-    b_base,
-    M,
-    H: tl.constexpr,
-    pid,
-    eps,
-    BLOCK_M: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    NORMK: tl.constexpr,
-    NSTAGES: tl.constexpr = 3,
-    SCALE_BLK: tl.constexpr = 1 << 30,
-):
-    """FP8-weight fused RMSNorm + GEMV: y[r] = sum_k ws[r,blk(k)]*W8[r,k]*(rmsnorm(x)*g)[k].
-
-    W is FP8-e4m3 [M,H]; ws is a weight scale of granularity SCALE_BLK along K (one
-    scale per [row, k-block of SCALE_BLK]). SCALE_BLK>=H is per-row (ws shape [M,1]);
-    SCALE_BLK=32 is MXFP8-style 1x32 blocks ([M, H/32]). Activation stays fp32; each
-    program recomputes the norm scale from the stable x_ptr (no barrier). BLOCK_K must
-    be a multiple of SCALE_BLK (or >=H for per-row)."""
-    noff = tl.arange(0, NORMK)
-    nmask = noff < H
-    xall = tl.load(x_ptr + noff, mask=nmask, other=0.0).to(tl.float32)
-    ss = tl.sum(xall * xall, axis=0)
-    rms = 1.0 / tl.sqrt(ss / H + eps)
-    npid = tl.num_programs(0)
-    ko = tl.max_contiguous(tl.multiple_of(tl.arange(0, BLOCK_K), BLOCK_K), BLOCK_K)
-    n_tiles = (M + BLOCK_M - 1) // BLOCK_M
-    NK: tl.constexpr = (H + BLOCK_K - 1) // BLOCK_K
-    PERROW: tl.constexpr = SCALE_BLK >= H
-    NSB: tl.constexpr = (H + SCALE_BLK - 1) // SCALE_BLK  # scale blocks along K (per matrix row)
-    SBT: tl.constexpr = BLOCK_K // SCALE_BLK if not PERROW else 1  # scale blocks per K-tile
-    tile = pid
-    while tile < n_tiles:
-        rows = tile * BLOCK_M + tl.arange(0, BLOCK_M)
-        rmask = rows < M
-        acc = tl.zeros((BLOCK_M, BLOCK_K), dtype=tl.float32)  # per-row path (no inner scale)
-        bacc = tl.zeros((BLOCK_M,), dtype=tl.float32)  # per-block path (scale folded in K-loop)
-        for ki in tl.range(0, NK, num_stages=NSTAGES):
-            kk = ki * BLOCK_K + ko
-            kmask = kk < H
-            xk = tl.load(x_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            gk = tl.load(g_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            nk = xk * rms * gk
-            w = tl.load(
-                w_base + rows[:, None] * H + kk[None, :], mask=rmask[:, None] & kmask[None, :], other=0.0
-            ).to(tl.float32)
-            if PERROW:
-                acc += w * nk[None, :]
-            else:
-                # per-block dequant: psum[m,blk] = sum_{k in blk} W8*nk; then *= scale[m,blk]
-                part = (w * nk[None, :]).reshape(BLOCK_M, SBT, SCALE_BLK)
-                psum = tl.sum(part, axis=2)  # [BLOCK_M, SBT]
-                sbk = ki * SBT + tl.arange(0, SBT)
-                sc = tl.load(
-                    ws_base + rows[:, None] * NSB + sbk[None, :],
-                    mask=rmask[:, None] & (sbk[None, :] < NSB),
-                    other=0.0,
-                ).to(tl.float32)
-                bacc += tl.sum(psum * sc, axis=1)  # [BLOCK_M]
-        if PERROW:
-            s = tl.sum(acc, axis=1)
-            s = s * tl.load(ws_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        else:
-            s = bacc
-        if has_bias:
-            s += tl.load(b_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        tl.store(y_ptr + rows, s, mask=rmask)
-        tile += npid
-
-
-@triton.jit
-def _gemv_fp8_tiled(
-    w_base,
-    ws_base,
-    x_ptr,
-    y_ptr,
-    has_bias,
-    b_base,
-    M,
-    K: tl.constexpr,
-    pid,
-    BLOCK_M: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    NSTAGES: tl.constexpr = 3,
-    SCALE_BLK: tl.constexpr = 1 << 30,
-):
-    """FP8-weight GEMV (no norm): y[r] = sum_k ws[r,blk(k)]*W8[r,k]*x[k] (+b[r]).
-    Activation stays fp32. SCALE_BLK>=K is per-row; SCALE_BLK=32 is MXFP8 1x32 blocks."""
-    npid = tl.num_programs(0)
-    ko = tl.max_contiguous(tl.multiple_of(tl.arange(0, BLOCK_K), BLOCK_K), BLOCK_K)
-    n_tiles = (M + BLOCK_M - 1) // BLOCK_M
-    NK: tl.constexpr = (K + BLOCK_K - 1) // BLOCK_K
-    PERROW: tl.constexpr = SCALE_BLK >= K
-    NSB: tl.constexpr = (K + SCALE_BLK - 1) // SCALE_BLK
-    SBT: tl.constexpr = BLOCK_K // SCALE_BLK if not PERROW else 1
-    tile = pid
-    while tile < n_tiles:
-        rows = tile * BLOCK_M + tl.arange(0, BLOCK_M)
-        rmask = rows < M
-        acc = tl.zeros((BLOCK_M, BLOCK_K), dtype=tl.float32)
-        bacc = tl.zeros((BLOCK_M,), dtype=tl.float32)
-        for ki in tl.range(0, NK, num_stages=NSTAGES):
-            kk = ki * BLOCK_K + ko
-            kmask = kk < K
-            x = tl.load(x_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            w = tl.load(
-                w_base + rows[:, None] * K + kk[None, :], mask=rmask[:, None] & kmask[None, :], other=0.0
-            ).to(tl.float32)
-            if PERROW:
-                acc += w * x[None, :]
-            else:
-                part = (w * x[None, :]).reshape(BLOCK_M, SBT, SCALE_BLK)
-                psum = tl.sum(part, axis=2)
-                sbk = ki * SBT + tl.arange(0, SBT)
-                sc = tl.load(
-                    ws_base + rows[:, None] * NSB + sbk[None, :],
-                    mask=rmask[:, None] & (sbk[None, :] < NSB),
-                    other=0.0,
-                ).to(tl.float32)
-                bacc += tl.sum(psum * sc, axis=1)
-        if PERROW:
-            s = tl.sum(acc, axis=1)
-            s = s * tl.load(ws_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        else:
-            s = bacc
-        if has_bias:
-            s += tl.load(b_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        tl.store(y_ptr + rows, s, mask=rmask)
-        tile += npid
-
-
-@triton.jit
-def _gemv_fp8_resid_rmsnorm(
-    w_base,
-    ws_base,
-    x_ptr,
-    o_ptr,
-    g_ptr,
-    y_ptr,
-    has_bias,
-    b_base,
-    M,
-    H: tl.constexpr,
-    pid,
-    eps,
-    BLOCK_M: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    NORMK: tl.constexpr,
-    NSTAGES: tl.constexpr = 3,
-    SCALE_BLK: tl.constexpr = 1 << 30,
-):
-    """FP8-weight fused residual + RMSNorm + GEMV: y = GEMV8(rmsnorm(x + o) * g) with
-    weight scale of granularity SCALE_BLK along K. SCALE_BLK>=H is per-row; 32 is
-    MXFP8. Activation stays fp32."""
-    noff = tl.arange(0, NORMK)
-    nmask = noff < H
-    xall = tl.load(x_ptr + noff, mask=nmask, other=0.0).to(tl.float32) + tl.load(
-        o_ptr + noff, mask=nmask, other=0.0
-    ).to(tl.float32)
-    ss = tl.sum(xall * xall, axis=0)
-    rms = 1.0 / tl.sqrt(ss / H + eps)
-    npid = tl.num_programs(0)
-    ko = tl.max_contiguous(tl.multiple_of(tl.arange(0, BLOCK_K), BLOCK_K), BLOCK_K)
-    n_tiles = (M + BLOCK_M - 1) // BLOCK_M
-    NK: tl.constexpr = (H + BLOCK_K - 1) // BLOCK_K
-    PERROW: tl.constexpr = SCALE_BLK >= H
-    NSB: tl.constexpr = (H + SCALE_BLK - 1) // SCALE_BLK
-    SBT: tl.constexpr = BLOCK_K // SCALE_BLK if not PERROW else 1
-    tile = pid
-    while tile < n_tiles:
-        rows = tile * BLOCK_M + tl.arange(0, BLOCK_M)
-        rmask = rows < M
-        acc = tl.zeros((BLOCK_M, BLOCK_K), dtype=tl.float32)
-        bacc = tl.zeros((BLOCK_M,), dtype=tl.float32)
-        for ki in tl.range(0, NK, num_stages=NSTAGES):
-            kk = ki * BLOCK_K + ko
-            kmask = kk < H
-            xk = tl.load(x_ptr + kk, mask=kmask, other=0.0).to(tl.float32) + tl.load(
-                o_ptr + kk, mask=kmask, other=0.0
-            ).to(tl.float32)
-            gk = tl.load(g_ptr + kk, mask=kmask, other=0.0).to(tl.float32)
-            nk = xk * rms * gk
-            w = tl.load(
-                w_base + rows[:, None] * H + kk[None, :], mask=rmask[:, None] & kmask[None, :], other=0.0
-            ).to(tl.float32)
-            if PERROW:
-                acc += w * nk[None, :]
-            else:
-                part = (w * nk[None, :]).reshape(BLOCK_M, SBT, SCALE_BLK)
-                psum = tl.sum(part, axis=2)
-                sbk = ki * SBT + tl.arange(0, SBT)
-                sc = tl.load(
-                    ws_base + rows[:, None] * NSB + sbk[None, :],
-                    mask=rmask[:, None] & (sbk[None, :] < NSB),
-                    other=0.0,
-                ).to(tl.float32)
-                bacc += tl.sum(psum * sc, axis=1)
-        if PERROW:
-            s = tl.sum(acc, axis=1)
-            s = s * tl.load(ws_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        else:
-            s = bacc
-        if has_bias:
-            s += tl.load(b_base + rows, mask=rmask, other=0.0).to(tl.float32)
-        tl.store(y_ptr + rows, s, mask=rmask)
-        tile += npid
-
-
-@triton.jit
-def _gemv_fp4(
-    blk_base,
-    scl_base,
-    x_ptr,
-    y_ptr,
-    b_base,
-    has_bias,
-    M,
-    K: tl.constexpr,
-    NB: tl.constexpr,
-    pid,
-    gate_w,
-    ACCUM: tl.constexpr,
-):
-    """FP4 weight GEMV. weight row r: blk_base + r*(K//2) bytes (K/2), scales
-    scl_base + r*NB. Rows strided by pid -> each row owned by one program, so
-    plain store / load-add-store (no atomics). ACCUM: y[r]+=gate_w*(s+b) else
-    y[r]=gate_w*(s+b). Processes one 32-wide scale block at a time."""
-    pos32 = tl.arange(0, 32)
-    byte_idx = pos32 // 2
-    hi = (pos32 % 2) == 1
-    r = pid
-    half = K // 2
-    while r < M:
-        acc = tl.zeros((32,), dtype=tl.float32)
-        kb = 0
-        while kb < NB:
-            raw = tl.load(blk_base + r * half + kb * 16 + byte_idx).to(tl.int32)
-            nib = tl.where(hi, (raw >> 4) & 0xF, raw & 0xF)
-            sign = (nib & 0x8) != 0
-            mag = _fp4_lut(nib & 0x7)
-            val = tl.where(sign, -mag, mag)
-            se = tl.load(scl_base + r * NB + kb).to(tl.int32)
-            sc = tl.where(se > 0, tl.exp2((se - 127).to(tl.float32)), 0.0)
-            xk = tl.load(x_ptr + kb * 32 + pos32).to(tl.float32)
-            acc += val * sc * xk
-            kb += 1
-        s = tl.sum(acc, axis=0)
-        if has_bias:
-            s += tl.load(b_base + r).to(tl.float32)
-        s = gate_w * s
-        if ACCUM:
-            s += tl.load(y_ptr + r).to(tl.float32)
-        tl.store(y_ptr + r, s)
-        r += _NWG
-
-
-@triton.jit
-def _quant_norm_fp8(
-    x_ptr, o_ptr, g_ptr, fp8_ptr, scl_ptr, H: tl.constexpr, NB: tl.constexpr, pid, eps, NORMK: tl.constexpr
-):
-    """Fused residual + RMSNorm + FP8-E4M3 quantization, per-32-element E8M0.
-
-    Reads the residual x and attention output o, applies RMSNorm(x + o) * g, and
-    quantizes to FP8 in one pass, so the experts' shared input never has to be
-    written to HBM as BF16 first. Each program computes the rms scalar from the full
-    (x + o), then quantizes its 32-element blocks."""
-    noff = tl.arange(0, NORMK)
-    nmask = noff < H
-    xall = tl.load(x_ptr + noff, mask=nmask, other=0.0).to(tl.float32) + tl.load(
-        o_ptr + noff, mask=nmask, other=0.0
-    ).to(tl.float32)
-    ss = tl.sum(xall * xall, axis=0)
-    rms = 1.0 / tl.sqrt(ss / H + eps)
-    pos32 = tl.arange(0, 32)
-    b = pid
-    while b < NB:
-        off = b * 32 + pos32
-        xv = tl.load(x_ptr + off).to(tl.float32) + tl.load(o_ptr + off).to(tl.float32)
-        gv = tl.load(g_ptr + off).to(tl.float32)
-        x = xv * rms * gv
-        amax = tl.max(tl.abs(x), axis=0)
-        target = amax / 448.0
-        u = target.to(tl.int32, bitcast=True)
-        raw = (u >> 23) & 0xFF
-        raw = raw + tl.where((u & 0x7FFFFF) != 0, 1, 0)
-        raw = tl.where(amax > 0.0, tl.minimum(tl.maximum(raw, 0), 255), 0)
-        sc = tl.where(raw > 0, tl.exp2((raw - 127).to(tl.float32)), 1.0)
-        tl.store(fp8_ptr + off, (x / sc).to(tl.float8e4nv))
-        tl.store(scl_ptr + b, raw.to(tl.uint8))
-        b += _NWG
-
-
-@triton.jit
-def _gemv_fp4_scaled(
-    blk_base,
-    scl_base,
-    afp8_ptr,
-    ascl_ptr,
-    y_ptr,
-    b_base,
-    has_bias,
-    N,
-    K: tl.constexpr,
-    NB: tl.constexpr,
-    pid,
-    gate_w,
-    ACCUM: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    MTILE: tl.constexpr,
-    NSTAGES: tl.constexpr = 3,
-    FINALIZE: tl.constexpr = False,
-    x_ptr=None,
-    o_ptr=None,
-):
-    """y[n] = gate_w*(sum_k W[n,k]*a[k] + b[n]) via native FP4xFP8 scaled MFMA
-    (tl.dot_scaled -> v_mfma_scale_f32_16x16x128_f8f6f4 on gfx950).
-    W FP4 e2m1 packed [N, K//2] (low nibble = even k), weight scales e8m0 [N, NB].
-    a FP8 e4m3 [K], act scales e8m0 [NB]. Output rows tiled BLOCK_N across programs.
-
-    The weight is the dot_scaled lhs ([BLOCK_N, K]) so its contiguous-K bytes
-    coalesce into wide loads; the single-token activation is the broadcast rhs
-    ([K, MTILE], only column 0 is real)."""
-    SB: tl.constexpr = BLOCK_K // 32
-    NK: tl.constexpr = (K + BLOCK_K - 1) // BLOCK_K
-    colsN = tl.arange(0, MTILE)
-    tile = pid
-    half = K // 2
-    n_tiles = (N + BLOCK_N - 1) // BLOCK_N
-    while tile < n_tiles:
-        n = tile * BLOCK_N + tl.arange(0, BLOCK_N)
-        nmask = n < N
-        acc = tl.zeros((BLOCK_N, MTILE), dtype=tl.float32)
-        # pipelined K-loop: tl.range(num_stages) overlaps the next tile's loads with
-        # the current dot, which the plain while-loop did not do
-        for ki in tl.range(0, NK, num_stages=NSTAGES):
-            k0 = ki * BLOCK_K
-            kk = k0 + tl.arange(0, BLOCK_K)
-            kp = (k0 // 2) + tl.arange(0, BLOCK_K // 2)
-            kpmask = kp < half
-            sb = (k0 // 32) + tl.arange(0, SB)
-            sbmask = sb < NB
-            # lhs = weight [BLOCK_N, BLOCK_K] e2m1 (packed K, contiguous -> coalesced)
-            w = tl.load(blk_base + n[:, None] * half + kp[None, :], mask=nmask[:, None] & kpmask[None, :], other=0).to(
-                tl.uint8
-            )
-            wscl = tl.load(scl_base + n[:, None] * NB + sb[None, :], mask=nmask[:, None] & sbmask[None, :], other=0)
-            # rhs = activation [BLOCK_K, MTILE] e4m3, only column 0 carries the token
-            a = tl.load(afp8_ptr + kk[:, None], mask=(colsN[None, :] == 0) & (kk[:, None] < K), other=0.0)
-            ascl = tl.load(ascl_ptr + sb[None, :], mask=sbmask[None, :], other=0)
-            ascl = tl.broadcast_to(ascl, (MTILE, SB))
-            acc = tl.dot_scaled(w, wscl, "e2m1", a, ascl, "e4m3", acc=acc, out_dtype=tl.float32)
-        y = tl.sum(tl.where(colsN[None, :] == 0, acc, 0.0), axis=1)  # [BLOCK_N]
-        if has_bias:
-            y += tl.load(b_base + n, mask=nmask, other=0.0).to(tl.float32)
-        y = gate_w * y
-        if ACCUM:
-            y += tl.load(y_ptr + n, mask=nmask, other=0.0).to(tl.float32)
-        tl.store(y_ptr + n, y, mask=nmask)
-        if FINALIZE:
-            # this program owns rows n of the MoE output; finalize the residual
-            # x[n] += o[n] + moe[n] here (same program wrote moe[n]) to fold the
-            # final residual into the down phase and drop its barrier
-            xv = tl.load(x_ptr + n, mask=nmask, other=0.0).to(tl.float32)
-            ov = tl.load(o_ptr + n, mask=nmask, other=0.0).to(tl.float32)
-            tl.store(x_ptr + n, xv + ov + y, mask=nmask)
-        tile += _NWG
 
 
 # ───────────────────────── the megakernel ─────────────────────────
@@ -836,78 +208,14 @@ def gpt_oss_megakernel(
         # for future tokens; the history [lo, pos-1] is read from the cache. Since no
         # attention head reads the just-written cache[pos], the append and the
         # attention can share a phase. ----
-        d = tl.arange(0, DH)
         # the KV-head owners append the current k,v (RoPE'd) to the cache for next time
         if pid < NKV:
-            h = tl.arange(0, HALF)
-            cosv = tl.load(cos_p + h).to(tl.float32)
-            sinv = tl.load(sin_p + h).to(tl.float32)
-            bidx = pid * DH
-            k1 = tl.load(k_p + bidx + h).to(tl.float32)
-            k2 = tl.load(k_p + bidx + HALF + h).to(tl.float32)
-            tl.store(kcache + pos * kv_dim + bidx + h, k1 * cosv - k2 * sinv)
-            tl.store(kcache + pos * kv_dim + bidx + HALF + h, k2 * cosv + k1 * sinv)
-            tl.store(vcache + pos * kv_dim + bidx + d, tl.load(v_p + bidx + d).to(tl.float32))
+            _rope_kv_append(k_p, v_p, cos_p, sin_p, kcache, vcache, pos, pid, kv_dim, DH, HALF)
         if pid < NH:
-            hh = pid
-            kvh = hh // GROUP
-            # RoPE this head's q in-register (NeoX half-split), full-width form:
-            #   d <  HALF: qr[d] = q[d]*cos[d]      - q[d+HALF]*sin[d]
-            #   d >= HALF: qr[d] = q[d]*cos[d-HALF] + q[d-HALF]*sin[d-HALF]
-            dlo = d % HALF  # angle index for both halves
-            cosd = tl.load(cos_p + dlo).to(tl.float32)
-            sind = tl.load(sin_p + dlo).to(tl.float32)
-            q_self = tl.load(q_p + hh * DH + d).to(tl.float32)
-            partner = tl.where(d < HALF, d + HALF, d - HALF)
-            q_part = tl.load(q_p + hh * DH + partner).to(tl.float32)
-            qv = tl.where(d < HALF, q_self * cosd - q_part * sind, q_self * cosd + q_part * sind)
-            lo = 0
-            if SLIDING > 0:
-                lo = tl.maximum(0, pos - SLIDING + 1)
-            # Flash-decode over the KV history [lo, pos-1] from the cache, online softmax.
-            m_i = -1e30
-            l_i = 0.0
-            acc = tl.zeros((DH,), dtype=tl.float32)
-            toff = tl.arange(0, BLOCK_T)
-            t0 = lo
-            while t0 < pos:
-                tt = t0 + toff
-                tmask = tt < pos
-                kblk = tl.load(
-                    kcache + tt[:, None] * kv_dim + kvh * DH + d[None, :], mask=tmask[:, None], other=0.0
-                ).to(tl.float32)
-                sc = tl.sum(kblk * qv[None, :], axis=1) * scale  # [BLOCK_T]
-                sc = tl.where(tmask, sc, -1e30)
-                blk_max = tl.max(sc, axis=0)
-                mn = tl.maximum(m_i, blk_max)
-                al = tl.exp(m_i - mn)
-                p = tl.exp(sc - mn)  # [BLOCK_T]
-                l_i = l_i * al + tl.sum(p, axis=0)
-                vblk = tl.load(
-                    vcache + tt[:, None] * kv_dim + kvh * DH + d[None, :], mask=tmask[:, None], other=0.0
-                ).to(tl.float32)
-                acc = acc * al + tl.sum(p[:, None] * vblk, axis=0)
-                m_i = mn
-                t0 += BLOCK_T
-            # current position pos: RoPE this kv-head's k in-register from k_p, which
-            # avoids reading the just-written cache entry and the append barrier.
-            kself = tl.load(k_p + kvh * DH + d).to(tl.float32)
-            kpart = tl.load(k_p + kvh * DH + partner).to(tl.float32)
-            k_pos = tl.where(d < HALF, kself * cosd - kpart * sind, kself * cosd + kpart * sind)
-            sc_pos = tl.sum(qv * k_pos, axis=0) * scale
-            mn = tl.maximum(m_i, sc_pos)
-            al = tl.exp(m_i - mn)
-            p_pos = tl.exp(sc_pos - mn)
-            l_i = l_i * al + p_pos
-            v_pos = tl.load(v_p + kvh * DH + d).to(tl.float32)
-            acc = acc * al + p_pos * v_pos
-            m_i = mn
-            sink = tl.load(sinks + hh).to(tl.float32)
-            mn = tl.maximum(m_i, sink)
-            al = tl.exp(m_i - mn)
-            l_i = l_i * al + tl.exp(sink - mn)
-            acc = acc * al / l_i
-            tl.store(attn_p + hh * DH + d, acc.to(tl.bfloat16))
+            _flash_decode_head(
+                q_p, k_p, v_p, cos_p, sin_p, kcache, vcache, sinks, attn_p,
+                pos, scale, pid, kv_dim, DH, HALF, GROUP, SLIDING, BLOCK_T,
+            )
         bc = 1  # barriers used so far this layer: QKV(1); KV-append fused into attention
         bc += 1
         _barrier(bar_p, (lb + bc) * _NWG)
@@ -949,20 +257,7 @@ def gpt_oss_megakernel(
         # in registers (E is tiny) and writes ids_p/gw_p. The writes are identical
         # across programs and each program reads back only what it wrote, so the
         # experts proceed without a separate top-k barrier.
-        eoff = tl.arange(0, E)
-        work = tl.load(logits_p + eoff).to(tl.float32)
-        for kk in range(0, TOPK):
-            mval = tl.max(work, axis=0)
-            ismax = work == mval
-            idx = tl.min(tl.where(ismax, eoff, E), axis=0)
-            tl.store(ids_p + kk, idx)
-            tl.store(gw_p + kk, mval)
-            work = tl.where(eoff == idx, -1e30, work)
-        tv = tl.load(gw_p + tl.arange(0, TOPK)).to(tl.float32)
-        tmax = tl.max(tv, axis=0)
-        ex = tl.exp(tv - tmax)
-        sm = tl.sum(ex, axis=0)
-        tl.store(gw_p + tl.arange(0, TOPK), ex / sm)
+        _topk_softmax(logits_p, ids_p, gw_p, E, TOPK)
 
         # ---- P6: experts. The top-k experts are independent until the final
         # accumulation, so each expert phase (gate-up, SwiGLU, down) runs over all
@@ -1005,39 +300,9 @@ def gpt_oss_megakernel(
         for slot in range(0, TOPK):
             gu_out = gu_p + slot * (2 * I)
             if QUANT:
-                afp8_out = afp8_p + slot * I
-                afp8_scl_out = afp8_scl_p + slot * DN_NB
-                pos32 = tl.arange(0, 32)
-                blk = pid
-                while blk < DN_NB:
-                    base_i = blk * 32 + pos32
-                    gate = tl.load(gu_out + 2 * base_i).to(tl.float32)
-                    up = tl.load(gu_out + 2 * base_i + 1).to(tl.float32)
-                    gate = tl.minimum(gate, limit)
-                    up = tl.maximum(tl.minimum(up, limit), -limit)
-                    glu = gate * (1.0 / (1.0 + tl.exp(-alpha * gate)))
-                    act = (up + 1.0) * glu  # [32]
-                    amax = tl.max(tl.abs(act), axis=0)
-                    target = amax / 448.0
-                    u = target.to(tl.int32, bitcast=True)
-                    raw = (u >> 23) & 0xFF
-                    raw = raw + tl.where((u & 0x7FFFFF) != 0, 1, 0)
-                    raw = tl.where(amax > 0.0, tl.minimum(tl.maximum(raw, 0), 255), 0)
-                    sc = tl.where(raw > 0, tl.exp2((raw - 127).to(tl.float32)), 1.0)
-                    tl.store(afp8_out + base_i, (act / sc).to(tl.float8e4nv))
-                    tl.store(afp8_scl_out + blk, raw.to(tl.uint8))
-                    blk += _NWG
+                _swiglu_quant_fp8(gu_out, afp8_p + slot * I, afp8_scl_p + slot * DN_NB, DN_NB, pid, alpha, limit)
             else:
-                act_out = act_p + slot * I
-                ii = pid
-                while ii < I:
-                    gate = tl.load(gu_out + 2 * ii).to(tl.float32)
-                    up = tl.load(gu_out + 2 * ii + 1).to(tl.float32)
-                    gate = tl.minimum(gate, limit)
-                    up = tl.maximum(tl.minimum(up, limit), -limit)
-                    glu = gate * (1.0 / (1.0 + tl.exp(-alpha * gate)))
-                    tl.store(act_out + ii, ((up + 1.0) * glu).to(tl.bfloat16))
-                    ii += _NWG
+                _swiglu_bf16(gu_out, act_p + slot * I, I, pid, alpha, limit)
         bc += 1
         _barrier(bar_p, (lb + bc) * _NWG)
         # --- phase C: down for every expert, accumulating gate-weighted sum into moe_p.
